Split step::serializer
