+++ ruststep/src/step/mod.rs
@@ -3,9 +3,11 @@
 mod logical;
 mod parameter;
 mod record;
+mod serializer;
 mod value;
 
 pub use logical::*;
 pub use parameter::*;
 pub use record::*;
+pub use serializer::*;
 pub use value::*;
