+++ ruststep/src/step/parameter.rs
@@ -1,6 +1,6 @@
 use crate::{error::*, step::*};
-use serde::{de, forward_to_deserialize_any, ser, Deserialize};
-use std::{convert::TryFrom, fmt};
+use serde::{de, forward_to_deserialize_any, Deserialize};
+use std::fmt;
 
 /// Primitive value type in STEP data
 ///
@@ -213,7 +213,7 @@ impl<'a> std::iter::FromIterator<&'a Parameter> for Parameter {
     }
 }
 
-impl<'de, 'param> de::Deserializer<'de> for &'param Parameter {
+impl<'de, 'se> de::Deserializer<'de> for &'se Parameter {
     type Error = Error;
 
     fn deserialize_any<V>(self, visitor: V) -> Result<V::Value>
@@ -260,7 +260,7 @@ impl<'de, 'param> de::Deserializer<'de> for &'param Parameter {
 }
 
 // To support SeqDeserializer
-impl<'de, 'param> de::IntoDeserializer<'de, crate::error::Error> for &'param Parameter {
+impl<'de, 'se> de::IntoDeserializer<'de, crate::error::Error> for &'se Parameter {
     type Deserializer = Self;
     fn into_deserializer(self) -> Self::Deserializer {
         self
@@ -323,273 +323,6 @@ impl<'de> de::Visitor<'de> for ParameterVisitor {
     {
         let (name, ty) = map.next_entry()?.unwrap();
         Ok(Parameter::Typed { name, ty })
-    }
-}
-
-impl<'param> ser::Serializer for &'param mut Parameter {
-    type Ok = ();
-    type Error = Error;
-    type SerializeSeq = Self;
-    type SerializeTuple = Self;
-    type SerializeTupleStruct = Self;
-    type SerializeTupleVariant = Self;
-    type SerializeMap = Self;
-    type SerializeStruct = Self;
-    type SerializeStructVariant = Self;
-
-    fn serialize_bool(self, v: bool) -> Result<()> {
-        todo!()
-    }
-
-    fn serialize_i8(self, v: i8) -> Result<()> {
-        self.serialize_i64(i64::from(v))
-    }
-    fn serialize_i16(self, v: i16) -> Result<()> {
-        self.serialize_i64(i64::from(v))
-    }
-    fn serialize_i32(self, v: i32) -> Result<()> {
-        self.serialize_i64(i64::from(v))
-    }
-    fn serialize_i64(self, v: i64) -> Result<()> {
-        todo!()
-    }
-
-    fn serialize_u8(self, v: u8) -> Result<()> {
-        self.serialize_i64(i64::from(v))
-    }
-    fn serialize_u16(self, v: u16) -> Result<()> {
-        self.serialize_i64(i64::from(v))
-    }
-    fn serialize_u32(self, v: u32) -> Result<()> {
-        self.serialize_i64(i64::from(v))
-    }
-    fn serialize_u64(self, v: u64) -> Result<()> {
-        self.serialize_i64(i64::try_from(v).expect("integer larger than i64::MAX is not supported"))
-    }
-
-    fn serialize_f32(self, v: f32) -> Result<()> {
-        self.serialize_f64(f64::from(v))
-    }
-    fn serialize_f64(self, v: f64) -> Result<()> {
-        todo!()
-    }
-
-    fn serialize_char(self, v: char) -> Result<()> {
-        self.serialize_str(&v.to_string())
-    }
-    fn serialize_str(self, v: &str) -> Result<()> {
-        todo!()
-    }
-    fn serialize_bytes(self, v: &[u8]) -> Result<()> {
-        todo!()
-    }
-
-    fn serialize_none(self) -> Result<()> {
-        self.serialize_unit()
-    }
-
-    fn serialize_some<T>(self, value: &T) -> Result<()>
-    where
-        T: ?Sized + ser::Serialize,
-    {
-        value.serialize(self)
-    }
-
-    fn serialize_unit(self) -> Result<()> {
-        todo!()
-    }
-
-    fn serialize_unit_struct(self, _name: &'static str) -> Result<()> {
-        self.serialize_unit()
-    }
-
-    fn serialize_unit_variant(
-        self,
-        _name: &'static str,
-        _variant_index: u32,
-        variant: &'static str,
-    ) -> Result<()> {
-        self.serialize_str(variant)
-    }
-
-    fn serialize_newtype_struct<T>(self, _name: &'static str, value: &T) -> Result<()>
-    where
-        T: ?Sized + ser::Serialize,
-    {
-        value.serialize(self)
-    }
-
-    fn serialize_newtype_variant<T>(
-        self,
-        _name: &'static str,
-        _variant_index: u32,
-        variant: &'static str,
-        value: &T,
-    ) -> Result<()>
-    where
-        T: ?Sized + ser::Serialize,
-    {
-        todo!()
-    }
-
-    fn serialize_seq(self, _len: Option<usize>) -> Result<Self::SerializeSeq> {
-        Ok(self)
-    }
-
-    fn serialize_tuple(self, len: usize) -> Result<Self::SerializeTuple> {
-        self.serialize_seq(Some(len))
-    }
-
-    fn serialize_tuple_struct(
-        self,
-        _name: &'static str,
-        len: usize,
-    ) -> Result<Self::SerializeTupleStruct> {
-        self.serialize_seq(Some(len))
-    }
-
-    fn serialize_tuple_variant(
-        self,
-        _name: &'static str,
-        _variant_index: u32,
-        variant: &'static str,
-        _len: usize,
-    ) -> Result<Self::SerializeTupleVariant> {
-        todo!()
-    }
-
-    fn serialize_map(self, _len: Option<usize>) -> Result<Self::SerializeMap> {
-        Ok(self)
-    }
-
-    fn serialize_struct(self, _name: &'static str, len: usize) -> Result<Self::SerializeStruct> {
-        self.serialize_map(Some(len))
-    }
-
-    fn serialize_struct_variant(
-        self,
-        _name: &'static str,
-        _variant_index: u32,
-        variant: &'static str,
-        _len: usize,
-    ) -> Result<Self::SerializeStructVariant> {
-        todo!()
-    }
-}
-
-impl<'param> ser::SerializeSeq for &'param mut Parameter {
-    type Ok = ();
-    type Error = Error;
-    fn serialize_element<T>(&mut self, value: &T) -> Result<()>
-    where
-        T: ?Sized + ser::Serialize,
-    {
-        todo!()
-    }
-    fn end(self) -> Result<()> {
-        todo!()
-    }
-}
-
-impl<'param> ser::SerializeTuple for &'param mut Parameter {
-    type Ok = ();
-    type Error = Error;
-    fn serialize_element<T>(&mut self, value: &T) -> Result<()>
-    where
-        T: ?Sized + ser::Serialize,
-    {
-        todo!()
-    }
-
-    fn end(self) -> Result<()> {
-        todo!()
-    }
-}
-
-impl<'param> ser::SerializeTupleStruct for &'param mut Parameter {
-    type Ok = ();
-    type Error = Error;
-
-    fn serialize_field<T>(&mut self, value: &T) -> Result<()>
-    where
-        T: ?Sized + ser::Serialize,
-    {
-        todo!()
-    }
-
-    fn end(self) -> Result<()> {
-        todo!()
-    }
-}
-
-impl<'param> ser::SerializeTupleVariant for &'param mut Parameter {
-    type Ok = ();
-    type Error = Error;
-
-    fn serialize_field<T>(&mut self, value: &T) -> Result<()>
-    where
-        T: ?Sized + ser::Serialize,
-    {
-        todo!()
-    }
-
-    fn end(self) -> Result<()> {
-        todo!()
-    }
-}
-
-impl<'param> ser::SerializeMap for &'param mut Parameter {
-    type Ok = ();
-    type Error = Error;
-
-    fn serialize_key<T>(&mut self, key: &T) -> Result<()>
-    where
-        T: ?Sized + ser::Serialize,
-    {
-        todo!()
-    }
-
-    fn serialize_value<T>(&mut self, value: &T) -> Result<()>
-    where
-        T: ?Sized + ser::Serialize,
-    {
-        todo!()
-    }
-
-    fn end(self) -> Result<()> {
-        todo!()
-    }
-}
-
-impl<'param> ser::SerializeStruct for &'param mut Parameter {
-    type Ok = ();
-    type Error = Error;
-
-    fn serialize_field<T>(&mut self, key: &'static str, value: &T) -> Result<()>
-    where
-        T: ?Sized + ser::Serialize,
-    {
-        todo!()
-    }
-
-    fn end(self) -> Result<()> {
-        todo!()
-    }
-}
-
-impl<'param> ser::SerializeStructVariant for &'param mut Parameter {
-    type Ok = ();
-    type Error = Error;
-
-    fn serialize_field<T>(&mut self, key: &'static str, value: &T) -> Result<()>
-    where
-        T: ?Sized + ser::Serialize,
-    {
-        todo!()
-    }
-
-    fn end(self) -> Result<()> {
-        todo!()
     }
 }
